requirements.txt
flask==3.0.3
openai==1.51.0
gunicorn==22.0.0
Werkzeug==3.0.0
python-dotenv==1.0.1
requests==2.26.0